resolve #33 로그인 시 노출되는 메뉴 변경, store 사용하도록 변경
