+++ search-where-to-go-web/src/store/constant.js
@@ -1,6 +1,5 @@
 export default {
   SHOW_ALERT_MODAL: 'showAlertModal',
   SET_USER: 'setUser',
-  LOGOUT: 'logout',
-  CHECK_AUTH: 'checkAuth'
+  LOGOUT: 'logout'
 }
